All file made changes to
I've added only the documents that I edited in some way to have the login/signup page working with comments marking parts that have to be transfared elsewhere once we have Db tables and a base main page. I will keep updating this branch right here whenever smth on my part changes and marking it with comments so its easier to read.